Fix L4 warnings.

+++ utils/scapp.h
@@ -569,7 +569,7 @@ scApp::processOption(argparser& parser, int opt)
             ktx_uint32_t level = strtoi(parser.optarg.c_str());
             level = clamp<ktx_uint32_t>(level, 0, KTX_PACK_UASTC_MAX_LEVEL);
             // Ensure the last one wins in case of multiple of these args.
-            options.bopts.uastcFlags = ~KTX_PACK_UASTC_LEVEL_MASK;
+            options.bopts.uastcFlags = (unsigned int)~KTX_PACK_UASTC_LEVEL_MASK;
             options.bopts.uastcFlags |= level;
             hasArg = true;
         }
